requirements.txt
requests>=2.31.0
InquirerPy>=0.3.4
rich>=13.0.0
pyyaml>=6.0.0
tiktoken>=0.5.0
